+++ ground/src/main/java/com/google/android/ground/ui/datacollection/GpsDataCollectionFragment.kt
@@ -9,13 +9,16 @@ import com.google.android.ground.databinding.GpsDataCollectionFragBinding
 import com.google.android.ground.model.task.Task
 import com.google.android.ground.ui.common.AbstractMapViewerFragment
 import com.google.android.ground.ui.editsubmission.AbstractTaskViewModel
+import com.google.android.ground.ui.home.mapcontainer.BaseMapViewModel
 import com.google.android.ground.ui.map.MapFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class GpsDataCollectionFragment
-constructor(private val task: Task, private val viewModel: AbstractTaskViewModel) :
-  AbstractMapViewerFragment() {
+class GpsDataCollectionFragment(
+  private val task: Task,
+  private val viewModel: AbstractTaskViewModel,
+  private val mapViewModel: BaseMapViewModel
+) : AbstractMapViewerFragment() {
 
   override fun onCreateView(
     inflater: LayoutInflater,
@@ -26,12 +29,15 @@ constructor(private val task: Task, private val viewModel: AbstractTaskViewModel
     val binding = GpsDataCollectionFragBinding.inflate(inflater, container, false)
 
     binding.lifecycleOwner = this
+    binding.mapViewModel = mapViewModel
     binding.setVariable(BR.viewModel, viewModel)
 
     return binding.root
   }
 
   override fun onMapReady(mapFragment: MapFragment) {
-//    mapFragment.disableGestures()
+    //    mapFragment.disableGestures()
   }
+
+  override fun getMapViewModel(): BaseMapViewModel = mapViewModel
 }
