+++ ground/src/main/java/com/google/android/ground/ui/offlinebasemap/selector/OfflineAreaSelectorFragment.kt
@@ -28,6 +28,7 @@ import com.google.android.ground.rx.RxAutoDispose.autoDisposable
 import com.google.android.ground.ui.common.AbstractMapViewerFragment
 import com.google.android.ground.ui.common.EphemeralPopups
 import com.google.android.ground.ui.common.Navigator
+import com.google.android.ground.ui.home.mapcontainer.BaseMapViewModel
 import com.google.android.ground.ui.map.MapFragment
 import com.google.android.ground.ui.map.gms.toGoogleMapsObject
 import com.google.android.ground.ui.offlinebasemap.selector.OfflineAreaSelectorViewModel.DownloadMessage
@@ -91,5 +92,9 @@ class OfflineAreaSelectorFragment : AbstractMapViewerFragment() {
       .map { it.toGoogleMapsObject() }
       .`as`(autoDisposable(this))
       .subscribe(viewModel::setViewport)
+  }
+
+  override fun getMapViewModel(): BaseMapViewModel {
+    TODO("Not yet implemented")
   }
 }
