+++ ground/src/main/res/layout/basemap.xml
@@ -1,0 +1,67 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:app="http://schemas.android.com/apk/res-auto"
+  xmlns:tools="http://schemas.android.com/tools">
+
+  <data>
+    <import type="androidx.core.content.ContextCompat" />
+    <import type="android.view.View" />
+    <variable
+      name="viewModel"
+      type="com.google.android.ground.ui.home.mapcontainer.BaseMapViewModel" />
+  </data>
+
+  <androidx.constraintlayout.widget.ConstraintLayout
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:background="@color/colorGrey800">
+
+    <!-- Placeholder for map implementation -->
+    <FrameLayout
+      android:id="@+id/map"
+      android:layout_width="match_parent"
+      android:layout_height="match_parent" />
+
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+      android:id="@+id/map_type_btn"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:onClick="@{() -> viewModel.onMapTypeButtonClicked()}"
+      android:tint="@color/colorGrey500"
+      app:backgroundTint="@color/colorBackground"
+      app:fabSize="mini"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintTop_toTopOf="parent"
+      app:srcCompat="@drawable/map_layers"
+      app:useCompatPadding="true" />
+
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+      android:id="@+id/location_lock_btn"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:enabled="@{viewModel.getLocationLockEnabled()}"
+      android:onClick="@{() -> viewModel.onLocationLockClick()}"
+      app:backgroundTint="@color/colorBackground"
+      app:fabSize="normal"
+      app:layout_constraintBottom_toTopOf="@+id/recycler_view"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:srcCompat="@drawable/ic_gps_lock"
+      app:tint="@{viewModel.getIconTint()}"
+      app:useCompatPadding="true" />
+
+    <!-- TODO: LOI cards are currently in the basemap layout and not home_layout because we want
+          it to be anchored to the GPS Lock button. -->
+    <androidx.recyclerview.widget.RecyclerView
+      android:id="@+id/recycler_view"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:layout_gravity="bottom"
+      android:orientation="horizontal"
+      app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintStart_toStartOf="parent"
+      tools:itemCount="3"
+      tools:listitem="@layout/loi_card_item" />
+  </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
