+++ ground/src/main/res/navigation/nav_graph.xml
@@ -87,9 +87,6 @@
     <action
       android:id="@+id/action_home_screen_fragment_to_locationOfInterestSelectorFragment"
       app:destination="@id/locationOfInterestSelectorFragment" />
-    <action
-      android:id="@+id/action_home_screen_fragment_to_mapTypeDialogFragment"
-      app:destination="@id/mapTypeDialogFragment" />
   </fragment>
   <fragment
     android:id="@+id/sync_status_fragment"
@@ -115,7 +112,7 @@
     android:id="@+id/question_data_collection_fragment"
     android:name="com.google.android.ground.ui.datacollection.QuestionDataCollectionFragment"
     android:label="@string/collect_data"
-    tools:layout="@layout/question_data_collection_frag"></fragment>
+    tools:layout="@layout/question_data_collection_frag" />
   <fragment
     android:id="@+id/offline_areas_fragment"
     android:name="com.google.android.ground.ui.offlinebasemap.OfflineAreasFragment"
@@ -222,6 +219,9 @@
     android:id="@+id/showHomeScreen"
     app:destination="@id/home_screen_fragment"
     app:popUpTo="@id/navGraph" />
+  <action
+    android:id="@+id/showMapTypeDialogFragment"
+    app:destination="@id/mapTypeDialogFragment" />
 
   <fragment
     android:id="@+id/terms_of_service_fragment"
